Revert "[Truffle] Add ImplicitCast from int to long." and related commits to check perf.
This reverts commit a483eac4ffd79122e4450198136a5abfee8288f2.
This reverts commit 375091582754a23f7d1ab464c03322689f64619b.
This reverts commit 2e4b534000e0c30440279734f32b1e2d70aa9413.

+++ core/src/main/java/org/jruby/truffle/nodes/core/MathNodes.java
@@ -543,12 +543,12 @@ public abstract class MathNodes {
 
         @Specialization
         public double function(int a, int b) {
-            return function((double) a, (double) b);
+            return function((double) a, b);
         }
 
         @Specialization
         public double function(int a, long b) {
-            return function((double) a, (double) b);
+            return function((double) a, b);
         }
 
         @Specialization
@@ -558,12 +558,12 @@ public abstract class MathNodes {
 
         @Specialization
         public double function(long a, int b) {
-            return function((double) a, (double) b);
+            return function((double) a, b);
         }
 
         @Specialization
         public double function(long a, long b) {
-            return function((double) a, (double) b);
+            return function((double) a, b);
         }
 
         @Specialization
@@ -573,17 +573,17 @@ public abstract class MathNodes {
 
         @Specialization
         public double function(BigInteger a, int b) {
-            return function(SlowPathBigInteger.doubleValue(a), (double) b);
+            return function(SlowPathBigInteger.doubleValue(a), b);
         }
 
         @Specialization
         public double function(BigInteger a, long b) {
-            return function(SlowPathBigInteger.doubleValue(a), (double) b);
+            return function(SlowPathBigInteger.doubleValue(a), b);
         }
 
         @Specialization
         public double function(BigInteger a, double b) {
-            return function(SlowPathBigInteger.doubleValue(a), (double) b);
+            return function(SlowPathBigInteger.doubleValue(a), b);
         }
 
         @Specialization
